+++ src/redux/hooks/useOrderbook.js
@@ -1,0 +1,6 @@
+import useRedux from './useRedux';
+import { orderbookActions } from '../actions';
+
+const useOrderbook = () => useRedux('orderbookReducer', orderbookActions);
+
+export default useOrderbook;